requirements.txt
-i https://pypi.org/simple
appdirs==1.4.4
certifi==2023.7.22; python_version >= '3.6'
environs==9.5.0
importlib-metadata==6.8.0; python_version >= '3.8'
marshmallow==3.20.1; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pyee==8.2.2
pyppeteer==1.0.2
python-dotenv==1.0.0; python_version >= '3.8'
python-slugify==8.0.1
text-unidecode==1.3
tqdm==4.66.1; python_version >= '3.7'
urllib3==1.26.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
websockets==10.4; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
